+++ meta_package_manager/managers/npm.py
@@ -30,11 +30,14 @@ import json
 from boltons.cacheutils import cachedproperty
 
 from ..base import PackageManager
+from ..platform import MACOS
 
 
 class NPM(PackageManager):
 
     cli_path = '/usr/local/bin/npm'
+
+    platforms = frozenset([MACOS])
 
     def get_version(self):
         return self.run([self.cli_path, '--version'])